+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
   now just optionaly returns one Translation. So
   `CompiledRegexp::find` returns Option now instead of a (possibly
   empty) Vec.
+- Fix a problem with the '.' quantifier in Context and Multipass
+  regular expressions.
 
 ### Fixed
 - Fixed the parsing of in alternation expressions `match` rules, i.e.